requirements.txt
Django>=5.2.0
mysqlclient>=2.2.0
mysql-connector-python>=8.0.0
gunicorn>=21.0.0
psycopg2-binary>=2.9.0
whitenoise>=6.5.0
python-decouple>=3.8
dj-database-url>=2.0.0
